feat(audit): hardening SQLite WAL/concurrency, resilienza Yahoo 429 con circuit-breaker, risk metrics, benchmark S&P500/FTSE MIB, smart rebalancer, bot Telegram interattivo, UI fintech dark + e2e tests
- Fix critici: NameError MarketDataService in portfolio, contratti REST settings/alerts, PRAGMA per-connessione (busy_timeout=10000)
- Nuovi endpoint: /health, /api/portfolio/risk-metrics, /benchmarks, /rebalance/*
- Bot Telegram bidirezionale: /value /radar /advice <TICKER>
- Frontend: marquee flash, chart area/candele+volumi+breakeven, benchmark overlay, rebalancer UI, skeleton/pulse, drawer mobile con swipe
- Test: tests/e2e_test.py (83 verifiche async)

+++ backend/models/portfolio.py
@@ -8,7 +8,7 @@ class Holding(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     stock_id = Column(Integer, ForeignKey("stocks.id"), nullable=False)
-    quantity = Column(Integer, nullable=False)
+    quantity = Column(Float, nullable=False)  # Float per supportare frazioni (ETF, rebalancer)
     avg_purchase_price = Column(Float, nullable=False)
     purchase_date = Column(Date, nullable=True)
     notes = Column(String, nullable=True)
